[WIP] a prototype of stacking NemotronH and EarTTS into a single model
This is an attempt to stack models together into a single engine.
It is required for speech 2 speech models which typically consists of
multiple sub-modules. Here we put EarTTS on top of NemotronH LLM.
One of the challenges that was uncovered: block size for attention
and mamba layers is adjusted to have same page size in bytes.
But since we have multiple types of attention layers (tts and llm),
this logic has to be expanded to compute custom block size per layer type.
Currently block size and page size padded are hardcoded in `gpu_model_runner`.
Stacking components together allows trace entire model into a single
cuda graph and avoid overhead of batching/unbatching per component.

Signed-off-by: Viacheslav Klimkov <vklimkov@nvidia.com>